backend/requirements.txt
# fastapi==0.109.0
# uvicorn[standard]==0.27.0
# python-multipart==0.0.6
# sqlalchemy==2.0.25
# psycopg2-binary==2.9.9
# python-dotenv==1.0.0
# pydantic==2.5.3
# pydantic-settings==2.1.0
# google-generativeai==0.3.2
# groq==0.9.0
# openai==1.10.0
# pillow==10.2.0
# pdf2image==1.17.0
# python-magic==0.4.27
# aiofiles==23.2.1
# httpx==0.26.0
# gtts==2.5.0


fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
google-generativeai==0.3.2
groq==0.9.0
openai==1.10.0
pillow==10.2.0
pdf2image==1.17.0
python-magic==0.4.27
aiofiles==23.2.1
httpx==0.24.1
gtts==2.5.0